Add generalized ListView component
