+++ frontend/src/components/mealrequest/ASPListView.tsx
@@ -1,15 +1,12 @@
 import { gql, useLazyQuery } from "@apollo/client";
 import {
   ChevronDownIcon,
-  ChevronLeftIcon,
-  ChevronRightIcon,
   ChevronUpIcon,
   DeleteIcon,
   EditIcon,
 } from "@chakra-ui/icons";
 import {
   Box,
-  Center,
   Button as ChakraButton,
   Collapse,
   Flex,
@@ -21,12 +18,6 @@ import {
   MenuOptionGroup,
   Text,
 } from "@chakra-ui/react";
-import {
-  DEFAULT_OPTIONS,
-  getTheme,
-} from "@table-library/react-table-library/chakra-ui";
-import { CompactTable } from "@table-library/react-table-library/compact";
-import { useTheme } from "@table-library/react-table-library/theme";
 import * as TABLE_LIBRARY_TYPES from "@table-library/react-table-library/types/table";
 import React, { useEffect, useState } from "react";
 import { BsFilter } from "react-icons/bs";
@@ -41,7 +32,7 @@ import {
 } from "../../types/MealRequestTypes";
 import { Contact } from "../../types/UserTypes";
 import { logPossibleGraphQLError } from "../../utils/GraphQLUtils";
-import LoadingSpinner from "../common/LoadingSpinner";
+import ListView from "../common/ListView";
 
 const GET_MEAL_REQUESTS_BY_ID = gql`
   query GetMealRequestsByRequestorId(
@@ -101,41 +92,9 @@ const GET_MEAL_REQUESTS_BY_ID = gql`
   }
 `;
 
-type ListViewProps = { authId: string; rowsPerPage?: number };
+type ASPListViewProps = { authId: string; rowsPerPage?: number };
 
-const ListView = ({ authId, rowsPerPage = 10 }: ListViewProps) => {
-  const chakraTheme = getTheme(DEFAULT_OPTIONS);
-  const customTheme = {
-    Table: `
-        margin: 0 !important;
-        width: 100%;
-        --data-table-library_grid-template-columns: repeat(4, minmax(0, 1fr)) 88px;
-  
-        .animate {
-          grid-column: 1 / -1;
-  
-          display: flex;
-        }
-  
-        .animate > div {
-          flex: 1;
-          display: flex;
-        }
-      `,
-    HeaderRow: `
-        background-color: var(--chakra-colors-gray-50);
-        color: var(--chakra-colors-gray-500);
-        font-family: Inter;
-        font-size: 14px;
-        font-style: normal;
-        font-weight: 600;
-        line-height: 21px;
-        text-transform: none;
-      `,
-  };
-
-  const theme = useTheme([chakraTheme, customTheme]);
-
+const ASPListView = ({ authId, rowsPerPage = 10 }: ASPListViewProps) => {
   const [ids, setIds] = React.useState<Array<TABLE_LIBRARY_TYPES.Identifier>>(
     [],
   );
@@ -162,7 +121,6 @@ const ListView = ({ authId, rowsPerPage = 10 }: ListViewProps) => {
     setCurrentlyEditingMealRequestId,
   ] = useState<string | undefined>(undefined);
 
-  // type TableNodeMealRequest = TABLE_LIBRARY_TYPES.TableNode & {};
   const [
     getMealRequests,
     {
@@ -369,20 +327,6 @@ const ListView = ({ authId, rowsPerPage = 10 }: ListViewProps) => {
     );
   }
 
-  if (getMealRequestsLoading || !data) {
-    return (
-      <Box
-        display="flex"
-        alignItems="center"
-        justifyContent="center"
-        w="100%"
-        h="200px"
-      >
-        <LoadingSpinner />
-      </Box>
-    );
-  }
-
   return (
     <>
       {currentlyEditingMealRequestId ? (
@@ -477,70 +421,18 @@ const ListView = ({ authId, rowsPerPage = 10 }: ListViewProps) => {
             </MenuList>
           </Menu>
         </Flex>
-        <Box
-          display="flex"
-          alignItems="center"
-          w="100%"
-          h="64px"
-          p="12px 16px"
-          bgColor="gray.50"
-          borderLeft="2px solid"
-          borderColor="gray.gray83"
-        >
-          <Text variant="desktop-body-bold">Meal Requests</Text>
-        </Box>
-        <CompactTable
+        <ListView
           columns={COLUMNS}
           rowOptions={ROW_OPTIONS}
           data={data}
-          theme={theme}
-          layout={{ custom: true }}
+          loading={getMealRequestsLoading}
+          requestType="Meal Requests"
+          currentPage={currentPage}
+          setCurrentPage={setCurrentPage}
         />
-        {getMealRequestsData?.getMealRequestsByRequestorId.length === 0 && (
-          <Center h="100px">
-            <Text>No meal requests to display</Text>
-          </Center>
-        )}
-        <Box
-          display="flex"
-          alignItems="center"
-          w="100%"
-          h="32px"
-          p="12px 16px"
-          bgColor="gray.50"
-          border="1px solid #E2E8F0"
-          borderRadius="0px 0px 8px 8px"
-          gap="16px"
-          color="#4A5568"
-          justifyContent="right"
-        >
-          <Text fontSize="14px">Page: {currentPage}</Text>
-          {currentPage === 1 ? (
-            <ChevronLeftIcon w="24px" h="24px" color="#A0AEC0" />
-          ) : (
-            <ChevronLeftIcon
-              w="24px"
-              h="24px"
-              cursor="pointer"
-              onClick={() => setCurrentPage(currentPage - 1)}
-            />
-          )}
-          {data?.nodes &&
-          data.nodes.length !== 0 &&
-          data.nodes.length % 5 === 0 ? (
-            <ChevronRightIcon
-              w="24px"
-              h="24px"
-              cursor="pointer"
-              onClick={() => setCurrentPage(currentPage + 1)}
-            />
-          ) : (
-            <ChevronRightIcon w="24px" h="24px" color="#A0AEC0" />
-          )}
-        </Box>
-      </Box>
+      </Box> 
     </>
   );
 };
 
-export default ListView;
+export default ASPListView;
